pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "pinterest-dl"
description = "An unofficial Pinterest image downloader"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "sean1832", email = "dev@zekezhang.com" }]
requires-python = ">=3.10"
keywords = ["pinterest", "image downloader", "web scraping"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Environment :: Console",
]
dependencies = [
    "playwright>=1.40.0",
    "selenium>=4.26.1",
    "pillow==10.4.0",
    "tqdm",
    "requests",
    "deprecated",
    "m3u8",
    "cryptography",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-mock>=3.10.0"]
exif = ["pyexiv2"]

[project.scripts]
pinterest-dl = "pinterest_dl.cli:main"
pin-dl = "pinterest_dl.cli:main"

[project.urls]
Homepage = "https://github.com/sean1832/pinterest-dl"
Issues = "https://github.com/sean1832/pinterest-dl/issues"

[tool.setuptools.dynamic]
version = { attr = "pinterest_dl.__version__" }

[tool.setuptools]
packages = { find = { include = ["pinterest_dl*"] } }
